Refs #2939

+++ src/cpp/rtps/participant/RTPSParticipantImpl.cpp
@@ -1103,7 +1103,7 @@ std::vector<std::string> RTPSParticipantImpl::getParticipantNames() const {
 void RTPSParticipantImpl::sendSync(CDRMessage_t* msg, Endpoint* /*pend*/, const Locator_t& destination_loc)
 {
     std::lock_guard<std::mutex> guard(m_send_resources_mutex);
-    for (int i = 0; i < m_senderResourceList.size(); ++i)
+    for (size_t i = 0; i < m_senderResourceList.size(); ++i)
     {
         auto& it = m_senderResourceList[i];
         bool sendThroughResource = false;
